tiktok-master/requirement.txt
astroid==2.2.5
Babel==2.7.0
Django==2.2.2
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
django-phone-field==1.7.1
django-phonenumber-field==3.0.1
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
phonenumbers==8.10.13
phonenumberslite==8.10.13
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.1

